pyproject.toml
[project]
name = "dawarich-cleaner"
version = "2.0.0"
description = "GPS outlier detection and cleaning for Dawarich"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.32.0",
    "httpx==0.27.2",
    "sqlalchemy[asyncio]==2.0.36",
    "aiosqlite==0.20.0",
    "asyncpg==0.30.0",
    "pytz==2024.2",
    "jinja2==3.1.4",
    "python-multipart==0.0.12",
    "apscheduler==3.10.4",
    "pydantic-settings==2.6.1",
    "python-dotenv==1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.4",
    "pytest-asyncio==0.24.0",
    "black==24.10.0",
    "ruff==0.8.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
